WU 23_conversions OK

+++ exercises/23_conversions/using_as.rs
@@ -1,3 +1,8 @@
+/// Write up
+/// ici on rajoute un "as f64" pour garantir que le résultat de la 
+/// division est du type f64.
+
+
 // using_as.rs
 //
 // Type casting in Rust is done via the usage of the `as` operator. Please note
@@ -10,11 +15,9 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
-
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
